Fixed resource query.
Now the resource query works. The extension methods are much simpler and simpler to use. However since the .net getmanifestresourceinfo gives null in namespace and filename a hack has been introduced.

You cannot have files with more than one dot in it and the ie. docker-compose.yaml works but not docker.compose.yaml. The extension length is max 4 characters otherwise it will treat the extension as the filename.