Added support for volumes.
Now volumes are discoverable and can be created & deleted. Added methods to get attached volumes on a container as well as networks that a container has attached to. A new volume builder to be used to create or ensure existence of containers and the container builder is updated to allow for mounting of those containers.