Added support for docker execute in fluent API.
The fluent API supports execution one or more binaries including arguments when container entered running state and just before shutting down.

Therefore it is possible to run commands needed to be run after a certain process or port is available and then to do e.g. cleanup or other when shutting down.