Added support for pause, unpause and fix Issue #112, 113.
This makes the composite service to use docker-compose as aligned with
the container service. However, the Dispose method must use the down
since it needs to handle remove images and that is not present in the
docker-compose rm API. It also adds KeepRunning and KeepContainer of
which replaces (but kept for backward compatability) KeepOnDispose. The
KeepOnDispose invokes internally KeepContainer. This commit also adds
pause & resume on ContainerService.