Use PublishAllPorts instead of explicit port mappings in unit tests
If your test runner is configured to run several tests in parallel you will end up with bind errors because the port 40001 is already in use by another test.
> Ductus.FluentDocker.Common.FluentDockerException: Failed to start container modest_ramanujan log: Error response from daemon: driver failed programming external connectivity on endpoint modest_ramanujan (3e80ba4d0e93156184a77c573fe0b743c8e52cd590f3d58e0e4b6c9cc65cca3c): Bind for 0.0.0.0:40001 failed: port is already allocated

Also remove `ExposePort(40001, 5432)` in `PauseAndResumeShallWorkOnSingleContainer` where port mapping doesn't matter for the the test but could interfere with the `ExplicitPortMappingShouldWork` test.