Parse process timespans with format 0h0m0s
This introduces a slightly more complex parsing method for timespans
in the ProcessRow class in order to deal with more diverse formats (such
as the format output by `podman top` - 1h2m3s vs 01:02:03).

This also adds a simple test to validate the parser's current behaviour
with docker, and a test to confirm it can parse podman output too
without crashing.